make return value long

+++ app/src/main/java/com/kot/bitcoinchart/data/entities/PriceValue.kt
@@ -15,6 +15,6 @@ For support, please feel free to contact me at https://www.linkedin.com/in/syeda
 
 data class Values (
 
-	@SerializedName("x") val x : Int,
+	@SerializedName("x") val x : Long,
 	@SerializedName("y") val y : Double
 )
